Split scope helpers
