+++ app/scripts/lib/multichain-api/scope/index.ts
@@ -1,0 +1,5 @@
+export * from './authorization'
+export * from './scope'
+export * from './supported'
+export * from './transform'
+export * from './validation'
